pyproject.toml
[project]
python = "^3.13"
name = "electronic-prescription-service-api-regression-tests"

[tool.poetry]
name = "electronic-prescription-service-api-regression-tests"
version = "0.1.0"
description = "Reusable test utilities for NHS Electronic Prescription Service (EPS) testing"
authors = ["SeanSteberis <sean.steberis@airelogic.com>"]
readme = "README.md"
packages = [{include = "eps_test_support", from = "src"}]

[[tool.poetry.source]]
name = "github-packages"
url = "https://pypi.pkg.github.com/NHSDigital/electronic-prescription-service-api-regression-tests/simple"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.13"
jsonschema = "^4.25.1"
requests = "^2.32.5"
assertpy = "^1.1"
behave = "1.2.6"
behave-cucumber-formatter = "^1.0.1"
allure-behave = "^2.15.3"
black = "^26.1.0"
semver = "^3.0.4"
gitpython = "^3.1.46"
pip-licenses = "^5.5.1"
flake8 = "^7.3.0"
pre-commit = "^4.5.1"
isort = "^7.0.0"
cryptography = "^46.0.4"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.2.1"
requests-oauthlib = "^2.0.0"
nhs-number-generator = {git = "https://github.com/Iain-S/nhs_number_generator.git"}
pyright = "^1.1.407"
playwright = "^1.58.0"
jycm = "^1.5.0"
pyotp = "^2.9.0"
lxml = "5.4.0"
pytest-nhsd-apim = "^6.0.6"
boto3 = "^1.42.38"
aiofiles = "^25.1.0"
wheel = "^0.46.2"
numpy = "^2.4.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.1"

[tool.black]
line-length = 120
target-version = ['py313']

[tool.pyright]
include = ["features", "src/eps_test_support", "scripts"]
exclude = ["**/.*",
    "**/__pycache__",
    "**/venv",
    "**/.venv",
    "**/node_modules"
]
defineConstant = { DEBUG = true }
reportMissingImports = true

pythonVersion = "3.13"
pythonPlatform = "Linux"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
